requirements.txt
requests
langchain
langchain_openai
python-dotenv
pandas